use_cases/sql_agents/my-sql-agent/pyproject.toml
[tool.poetry]
name = "my-sql-agent"
version = "0.1.0"
description = "An AI-powered SQL query generation and database interaction agent"
authors = ["tcashel <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.0"
langchain-core = "^0.1.0"
langchain-community = "^0.0.10"
langchain-openai = "^0.0.1"
langchain-experimental = "^0.0.49"
langchain-ollama = "^0.0.1"
langgraph = "^0.0.20"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.0"
pydantic = "^2.0.0"
httpx = "^0.24.0"
anyio = "^4.0.0"
openlit = "^1.33.8"
tomli = "^2.0.1"
shared = {path = "../../../shared", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.7.0"
ruff = "^0.1.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]